+++ packages/cli/src/server/install/Install.handler.ts
@@ -35,6 +35,7 @@ export default class ServerInstallHandler implements ICommandHandler {
         let caughtError: Error;
         try {
             const installStatus = await ZSshUtils.installServer(session, serverPath, {
+                identityAgent: params.arguments.identityAgent,
                 onProgress: (progressIncrement) => {
                     task.percentComplete += progressIncrement;
                 },